Fix source transpilation by making rewire the last plugin.
Forcing the plugin to be last is done by loading the plugin from a
preset instead (those run after plugins) and making it the first preset
(because presets are ran in opposite order).

Now there’s an issue with the code rewire generates which uses ES2015
exports, instead of ES5.